Created add Comment Modal, add Comment button, view Comment Button
I added the add comment button and view comment button. When the add comment button is clicked it toggles the Comment modal in the CreateComment.js file. When the modal is filled in it pushes the new comment into the correct place in the database. I still need to make sure that the user cannot put in nothing to the comment and submit it. I also need to work activating the view comment button.